Added files via upload
src1.go is the main program file, data.go does not contain code but it shows how the data should be stored on variable 'data'.(which is same as the regional code provided in question). It will give the first set of wrong data for all the distributor, because it may lead to inconsistency when we try to follow the underlying distributor based upon the distributor above whenever the program gives the first record of wrong data for a particular distributor, remove it to find the next record of wrong data. INPUT DATA is given in the variable of the program for simplicity and being not allowed to use any database.