User Permissions
This commit is all about the permissions and chekcing the possiblity of permission inheritence from parent user.
Some of the sample users are saved in the data.go in a hashmap and are used in the program.

The user with id 3 is an invalid user as the parent user 2 does not has permissions for locations in base.

how to run the program
----------------------

Build the main.go file and run it.
The instructions will come to select a user and check the permissions in areas.
If user does not satisfies permissions as per its parent, then the program will abort .
To change user, run the program again and select a different user id.